Change devops revert, create base test class
